+++ src/main/g8/default.properties
@@ -1,6 +1,7 @@
 name=zio-quickstart
 description=This is a seed project that creates Scala 3 based ZIO application using zio-http, protoquill, caliban etc.
 dotty_version=3.0.2
+add_http_endpoints_and_database_repositories=yes
 add_websocket_endpoint=yes
 add_graphql=yes
 add_metrics=yes
