+++ src/main/g8/src/main/resources/$if(add_http_endpoints_and_database_repositories.truthy)$graphql$endif$/graphiql.html
