+++ src/main/g8/src/main/resources/application.conf
@@ -10,7 +10,8 @@ diagnostics-server-config {
 }
 $endif$
 
-testPostgresDB {
+$if(add_http_endpoints_and_database_repositories.truthy)$
+postgres-db {
     dataSource {
         user=postgres
         databaseName=quickstart-db
@@ -20,3 +21,4 @@ testPostgresDB {
     }
     dataSourceClassName=org.postgresql.ds.PGSimpleDataSource
 }
+$endif$
