+++ src/main/g8/src/main/scala/$package$/$if(add_http_endpoints_and_database_repositories.truthy)$service$endif$/ItemServiceLive.scala
