+++ src/main/g8/src/main/scala/$package$/healthcheck/Healthcheck.scala
@@ -1,0 +1,12 @@
+package $package$.healthcheck
+
+import zio._
+import zhttp.http._
+
+object Healthcheck:
+
+  //TODO return HeathCheckStatus
+  val expose: HttpApp[Any, Throwable] = HttpApp.collectM {
+    case Method.GET -> Root / "health" =>
+        ZIO.succeed(Response.status(Status.OK))
+  }
