--- src/main/g8/src/main/scala/$package$/service/$if(add_websocket_endpoint.truthy)$SubscriberService.scala$endif$
@@ -1,20 +1,0 @@
-package $package$.service
-
-import $package$.domain.ItemId
-import zio._
-import zio.stream._
-
-trait SubscriberService:
-  def publishDeleteEvents(deletedItemId: ItemId): IO[Nothing, List[Boolean]]
-  def showDeleteEvents: Stream[Nothing, ItemId]
-  def getQueue: UIO[List[Queue[ItemId]]]
-  
-object SubscriberService:
-  def publishDeleteEvents(deletedItemId: ItemId): ZIO[Has[SubscriberService], Nothing, List[Boolean]] =
-    ZIO.serviceWith[SubscriberService](_.publishDeleteEvents(deletedItemId))
-
-  def showDeleteEvents: ZStream[Has[SubscriberService], Nothing, ItemId] =
-    ZStream.accessStream(_.get.showDeleteEvents)
-
-  def getQueue: ZIO[Has[SubscriberService], Nothing, List[Queue[ItemId]]] =
-    ZIO.serviceWith[SubscriberService](_.getQueue)
