+++ src/main/g8/src/test/$if(add_http_endpoints_and_database_repositories.truthy)$resources$endif$/item_schema.sql
