+++ src/main/g8/src/test/scala/$package$/$if(add_http_endpoints_and_database_repositories.truthy)$api$endif$/$if(add_graphql.truthy)$graphql$endif$/GraphqlQueries.scala
