+++ src/main/g8/src/test/scala/$package$/healthcheck/HealthCheckSpec.scala
@@ -1,0 +1,6 @@
+package $package$.healthcheck
+
+import zio.test._
+
+object HealthcheckSpec extends DefaultRunnableSpec:
+  def spec = ???
